# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Subscriptions Service API
  version: "1.0.0"
paths:
  /api/v1/subscriptions:
    post:
      summary: Create subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
    get:
      summary: List subscriptions
      parameters:
        - in: query
          name: user_id
          schema:
            type: string
            format: uuid
        - in: query
          name: service_name
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: offset
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Subscription'
  /api/v1/subscriptions/{id}:
    get:
      summary: Get subscription
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        '404':
          description: Not found
    put:
      summary: Update subscription
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionUpdateRequest'
      responses:
        '200':
          description: OK
    delete:
      summary: Delete subscription
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No content
  /api/v1/subscriptions/summary:
    get:
      summary: Calculate total price for period
      parameters:
        - in: query
          name: from
          required: true
          schema:
            type: string
            example: "07-2025"
        - in: query
          name: to
          required: true
          schema:
            type: string
            example: "10-2025"
        - in: query
          name: user_id
          schema:
            type: string
            format: uuid
        - in: query
          name: service_name
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer

components:
  schemas:
    SubscriptionCreateRequest:
      type: object
      required: [service_name, price, user_id, start_date]
      properties:
        service_name:
          type: string
        price:
          type: integer
        user_id:
          type: string
          format: uuid
        start_date:
          type: string
          example: "07-2025"
        end_date:
          type: string
          nullable: true
          example: "10-2025"
    SubscriptionUpdateRequest:
      type: object
      properties:
        service_name:
          type: string
        price:
          type: integer
        start_date:
          type: string
        end_date:
          type: string
          nullable: true
    Subscription:
      type: object
      properties:
        id:
          type: string
          format: uuid
        service_name:
          type: string
        price:
          type: integer
        user_id:
          type: string
          format: uuid
        start_date:
          type: string
        end_date:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time